+++ Assets/Scripts/Global/Services/Timer/TimerService.cs
@@ -12,11 +12,11 @@ namespace Global.Services.Timer {
             _timersToDelete = new HashSet<string>();
         }
 
-        public void StartTimer(string timerId, float time, Action onEnd, bool isLoop = false, Action<int> onTick = null) {
+        public void StartTimer(string timerId, float time, Action onEnd, bool isLoop = false, Action<float> onTick = null) {
             StartDownTimer(timerId, time, onEnd, isLoop, onTick);
         }
 
-        public void StartDownTimer(string timerId, float time, Action onEnd, bool isLoop = false, Action<int> onTick = null) {
+        public void StartDownTimer(string timerId, float time, Action onEnd, bool isLoop = false, Action<float> onTick = null) {
             var timer = new DownTimer(time, onEnd, isLoop, onTick);
             
             _timers.Add(timerId, timer);
@@ -28,7 +28,7 @@ namespace Global.Services.Timer {
             return _timers[timerId].GetTime();
         }
 
-        public void StartUpTimer(string timerId, float time, Action onEnd, bool isLoop = false, Action<int> onTick = null)
+        public void StartUpTimer(string timerId, float time, Action onEnd, bool isLoop = false, Action<float> onTick = null)
         {
             var timer = new UpTimer(time, onEnd, isLoop, onTick);
             _timers.Add(timerId, timer);
